backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pyodbc>=5.1.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
phonenumbers>=8.13.0
apscheduler>=3.10.0
redis>=5.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
openpyxl>=3.1.0
